+++ src/main/java/ru/kata/spring/boot_security/demo/service/UserService.java
@@ -9,7 +9,7 @@ public interface UserService {
     List<User> getUserList();
     void add(User user);
     void edit(int id, User user);
-    User getUserById(int id);
+    User getUserById(Long id);
     User findByEmail(String email);
-    void delete(int id);
+    void delete(Long id);
 }
