+++ src/main/java/ru/kata/spring/boot_security/demo/service/UserServiceImpl.java
@@ -45,7 +45,7 @@ public class UserServiceImpl implements UserService {
 
     @Override
     @Transactional(readOnly = true)
-    public User getUserById(int id) {
+    public User getUserById(Long id) {
         return userRepository.findById(id).orElseThrow();
 //        return userDao.getUserById(id);
     }
@@ -58,7 +58,7 @@ public class UserServiceImpl implements UserService {
 
     @Override
     @Transactional
-    public void delete(int id) {
+    public void delete(Long id) {
         userRepository.deleteById(id);
 //        userDao.delete(id);
     }
